+++ src/Skeletest/Internal/TestRunner.hs
@@ -59,6 +59,8 @@ class (MonadIO m) => Testable m where
 
 {----- TestResult -----}
 
+-- TODO: Remove 'testResult' prefix
+-- https://github.com/brandonchinn178/skeletest/issues/89
 data TestResult = TestResult
   { testResultSuccess :: Bool
   , testResultLabel :: Text