+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased
 
+Runtime changes:
+* Sanitize a literal "\`\`\`" line in snapshots ([#27](https://github.com/brandonchinn178/skeletest/issues/27))
+* Show diff when snapshot doesn't exist also ([#51](https://github.com/brandonchinn178/skeletest/issues/51))
+* Only update snapshots if test passed ([#25](https://github.com/brandonchinn178/skeletest/issues/25))
+
 ## v0.3.7
 
 New features: